Updated readme.md
CHANGED:
But wait, not that again, HTML in code, maybe we can take out it, but it should be in same template file. Maybe we will
make some annotation which will be replaced with HTML (with some unique id) from our template?

TO:
But wait, not that again, HTML in code, maybe we can take it out, but it should be in same template file. Maybe we will
make some annotation which will be replaced with HTML (with some unique id) from our template?